Some norm and correct error when || only
