+++ Sources/Execution/binary_exec.c
@@ -12,51 +12,7 @@
 
 #include "minishell.h"
 
-extern long long g_status;
-
-char	*get_exec(char *cmd)
-{
-	size_t	j;
-	char	*res;
-
-	j = 0;
-	while (cmd[j])
-		j++;
-	while (j != 0 && cmd[j] != '/')
-		j--;
-	j++;
-	res = ft_substr(cmd, j, ft_strlen(cmd));
-	return (res);
-}
-
-char	*get_path(char *cmd)
-{
-	size_t	i;
-	char	*res;
-
-	i = 0;
-	while (cmd[i])
-		i++;
-	while (cmd[i] != '/')
-		i--;
-	i++;
-	res = ft_substr(cmd, 0, i);
-	return (res);
-}
-
-void	manage_fd(int fd_in, int fd_out)
-{
-	if (fd_out != 1)
-	{
-		dup2(fd_out, STDOUT_FILENO);
-		close(fd_out);
-	}
-	if (fd_in != 0)
-	{
-		dup2(fd_in, STDIN_FILENO);
-		close(fd_in);
-	}
-}
+extern long long	g_status;
 
 void	binary_with_path(t_mshell *shell, char **cmd_arr)
 {
@@ -76,7 +32,7 @@ void	binary_with_path(t_mshell *shell, char **cmd_arr)
 		return (free(exec), free_arr(exec_split), free_struct(shell), exit(1));
 	temp = ft_strjoin(path, exec);
 	if (!temp)
-		return (free(exec), free_arr(exec_split), free_struct(shell), 
+		return (free(exec), free_arr(exec_split), free_struct(shell),
 			free(path), exit(EXIT_FAILURE));
 	free(path);
 	if (!check_access(temp))
